refactor,tests: displacements
